+++ docs_markdown/changelog.md
@@ -2,8 +2,13 @@
 
 # Changelog
 
+0.7.3
+
+* Fixed SSEChannelContainer.add and rm wrong exception thrown on no existent channel
+
 0.7.2
-\* Invalid Channel or Listener Exception now yields an error message to client (no logs anymore)
+
+* Invalid Channel or Listener Exception now yields an error message to client (no logs anymore)
 
 0.7.0